ROM: Implement o_read_addr_ready with an internal register.
Expand the interface by o_invalid_read_width in case that i_read_width is none of the defined enu_memory_access_width
enumeration values (can only happen in synthesized design).
In simulation the VHDL should prevent that.

The test program does not finish. There is now a new bug in the design.